# HskApi/Models/TodoItem.cs
namespace HskApi.Models
{
    public class TodoItem
    {
        public long Id { get; set; }
        public string? Name { get; set; }
        public bool IsComplete { get; set; }
        public int test1 { get; set; }
    }
}